Fix data type of nside when reading V2 and older healpix maps.
This fixes a crash on deserialization when reading older healpix maps in pipeline
info frames and data corruption in other cases. size_t is generally 64-bit, so using
it would create an off-by-4-bytes issue after the object header.

+++ maps/src/HealpixSkyMap.cxx
@@ -244,7 +244,7 @@ HealpixSkyMap::load(A &ar, unsigned v)
 {
 	using namespace cereal;
 	int store;
-	size_t nside;
+	uint32_t nside;
 	bool nested, shifted;
 
 	G3_CHECK_VERSION(v);
